Fixe byte swapping error in double.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@3091 72102866-910b-0410-8b05-ffd578937521

+++ Headers/gnustep/base/NSByteOrder.h
@@ -228,7 +228,7 @@ static inline unsigned long long
 NSSwapLongLong(unsigned long long in)
 {
     union swap {
-	unsigned long	num;
+	unsigned long long	num;
 	unsigned char	byt[8];
     } dst;
     union swap	*src = (union swap*)&in;
